back atualizado

+++ src/main/java/com/queueup/queueup/utils/FilaCliente.java
@@ -3,13 +3,13 @@ package com.queueup.queueup.utils;
 
 public class FilaCliente {
 	private int idCliente;
-	private int idRestaurante;
+	private Long idRestaurante;
 	private int posicao;
 	
 	public FilaCliente() {
 	}
 	
-	public FilaCliente(int idCliente, int idRestaurante, int posicao) {
+	public FilaCliente(int idCliente, Long idRestaurante, int posicao) {
 		this.idCliente = idCliente;
 		this.idRestaurante = idRestaurante;
 		this.posicao = posicao;
@@ -26,12 +26,12 @@ public class FilaCliente {
 	}
 
 
-	public int getIdRestaurante() {
+	public Long getIdRestaurante() {
 		return idRestaurante;
 	}
 
 
-	public void setIdRestaurante(int idRestaurante) {
+	public void setIdRestaurante(Long idRestaurante) {
 		this.idRestaurante = idRestaurante;
 	}
